test(charts): move PNG smoke test where it can actually run
The ultros crate's test binaries don't link on Windows and CI's test step is disabled, so the item_card smoke test was dead code. Rasterization coverage now lives in ultros-charts as an integration test with a resvg dev-dependency.
