+++ ultros/src/web/item_card.rs
@@ -77,61 +77,3 @@ pub(crate) async fn item_card(
         .header(header::CONTENT_TYPE, mime_type.as_ref())
         .body(Body::new(http_body_util::Full::from(bytes)))?)
 }
-
-#[cfg(test)]
-mod tests {
-    use super::svg_to_png;
-    use chrono::DateTime;
-    use ultros_api_types::SaleHistory;
-    use ultros_api_types::world::{Datacenter, Region, World, WorldData};
-    use ultros_api_types::world_helper::WorldHelper;
-    use ultros_charts::charts::price_history::{PriceChartOptions, build_price_history_scene};
-    use ultros_charts::svg::scene_to_svg;
-
-    #[test]
-    fn renders_a_decodable_png() {
-        let helper = WorldHelper::new(WorldData {
-            regions: vec![Region {
-                id: 1,
-                name: "Test".to_string(),
-                datacenters: vec![Datacenter {
-                    id: 1,
-                    name: "DC".to_string(),
-                    region_id: 1,
-                    worlds: vec![World {
-                        id: 1,
-                        name: "World".to_string(),
-                        datacenter_id: 1,
-                    }],
-                }],
-            }],
-        });
-        let sales: Vec<SaleHistory> = (0..30)
-            .map(|i| SaleHistory {
-                id: i,
-                quantity: 1,
-                price_per_item: 1_000 + i * 13,
-                buying_character_id: 0,
-                hq: false,
-                sold_item_id: 1,
-                sold_date: DateTime::from_timestamp(1_750_000_000 + i as i64 * 7_200, 0)
-                    .unwrap()
-                    .naive_utc(),
-                world_id: 1,
-                buyer_name: None,
-            })
-            .collect();
-        let scene = build_price_history_scene(
-            &helper,
-            &sales,
-            &PriceChartOptions {
-                title: Some("Smoke Test - Sale History".to_string()),
-                remove_outliers: true,
-                ..Default::default()
-            },
-        );
-        let png = svg_to_png(&scene_to_svg(&scene)).expect("png");
-        let decoded = image::load_from_memory(&png).expect("decodable png");
-        assert_eq!((decoded.width(), decoded.height()), (960, 540));
-    }
-}
